requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
openai>=1.0.0
easyocr>=1.7.0
numpy>=1.24.0
Pillow>=10.0.0

# Recommendation System
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
